Forget the previous commit! :-)
The serializers of those models are done but they probably need to be edited.

As it is now when you create a volunteer or an organization the relative Focus model doesn't get created. This still needs to be implemented.

The last point is not yet done:
`/api/calls/<int:vol_id>/` **GET**: Get the list of all the calls a volunteer sent a request to.